+++ src/main/java/jnum/jnumops/MatMulOps.java
@@ -37,9 +37,9 @@ public class MatMulOps {
 
 
     public static NDArray matmulFloat(NDArray a, NDArray b, NDArray resArray) {
-        int n = a.internalShapeUnsafe()[0]; 
-        int m = a.internalShapeUnsafe()[1]; 
-        int p = b.internalShapeUnsafe()[1];
+        int n = (int) a.internalShapeUnsafe()[0]; 
+        int m = (int) a.internalShapeUnsafe()[1]; 
+        int p = (int) b.internalShapeUnsafe()[1];
         
         try (Arena arena = Arena.ofShared()) {
             MemorySegment memA = a.isContiguous() ? a.getData() : a.contiguous(arena).getData();
@@ -620,9 +620,9 @@ public class MatMulOps {
 
 
     public static NDArray matmulDouble(NDArray a, NDArray b, NDArray resArray) {
-        int n = a.internalShapeUnsafe()[0]; 
-        int m = a.internalShapeUnsafe()[1]; 
-        int p = b.internalShapeUnsafe()[1];
+        int n = (int) a.internalShapeUnsafe()[0]; 
+        int m = (int) a.internalShapeUnsafe()[1]; 
+        int p = (int) b.internalShapeUnsafe()[1];
         
         try (Arena arena = Arena.ofShared()) {
             MemorySegment memA = a.isContiguous() ? a.getData() : a.contiguous(arena).getData();
@@ -1203,9 +1203,9 @@ public class MatMulOps {
 
 
     public static NDArray matmulInt(NDArray a, NDArray b, NDArray resArray) {
-        int n = a.internalShapeUnsafe()[0]; 
-        int m = a.internalShapeUnsafe()[1]; 
-        int p = b.internalShapeUnsafe()[1];
+        int n = (int) a.internalShapeUnsafe()[0]; 
+        int m = (int) a.internalShapeUnsafe()[1]; 
+        int p = (int) b.internalShapeUnsafe()[1];
         
         try (Arena arena = Arena.ofShared()) {
             MemorySegment memA = a.isContiguous() ? a.getData() : a.contiguous(arena).getData();
