+++ src/main/java/jnum/jnumops/ReduceOps.java
@@ -103,7 +103,7 @@ public class ReduceOps {
     }
 
     public static NDArray sumFloatAxis(NDArray a,int axis,NDArray resArray){
-        int size= a.internalShapeUnsafe()[axis];
+        long size= a.internalShapeUnsafe()[axis];
         long strideA= a.internalStridesUnsafe()[axis];
 
         for(int i=0;i<resArray.getSize();i++){
@@ -121,8 +121,8 @@ public class ReduceOps {
 
             if(strideA==1){
                 var vAcc = FloatVector.zero(SPECIES);
-                int k = 0;
-                int loopbound = SPECIES.loopBound(size);
+                long k = 0;
+                long loopbound = SPECIES.loopBound(size);
                 for(; k<loopbound ;k += VL){
                     var vVal = FloatVector.fromMemorySegment(SPECIES, a.getData(), (baseOffset + k) * 4L, ORDER);
                     vAcc = vAcc.add(vVal);
@@ -134,7 +134,7 @@ public class ReduceOps {
                 resArray.getData().setAtIndex(ValueLayout.JAVA_FLOAT, offsetRes, acc);
             }else{
                 float acc = 0f;
-                for (int k = 0; k < size; k++) {
+                for (long k = 0; k < size; k++) {
                     acc += a.getData().getAtIndex(ValueLayout.JAVA_FLOAT, baseOffset + k * strideA);
                 }
                 resArray.getData().setAtIndex(ValueLayout.JAVA_FLOAT, offsetRes, acc);
@@ -144,7 +144,7 @@ public class ReduceOps {
     }
 
     public static NDArray sumIntAxis(NDArray a,int axis,NDArray resArray){
-        int size= a.internalShapeUnsafe()[axis];
+        long size= a.internalShapeUnsafe()[axis];
         long strideA= a.internalStridesUnsafe()[axis];
 
         for(int i=0;i<resArray.getSize();i++){
@@ -162,8 +162,8 @@ public class ReduceOps {
 
             if(strideA==1){
                 var vAcc = IntVector.zero(SPECIESINT);
-                int k = 0;
-                int loopbound = SPECIESINT.loopBound(size);
+                long k = 0;
+                long loopbound = SPECIESINT.loopBound(size);
                 for(; k<loopbound ;k += INT_VL){
                     var vVal = IntVector.fromMemorySegment(SPECIESINT, a.getData(), (baseOffset + k) * 4L, ORDER);
                     vAcc = vAcc.add(vVal);
@@ -175,7 +175,7 @@ public class ReduceOps {
                 resArray.getData().setAtIndex(ValueLayout.JAVA_INT, offsetRes, acc);
             }else{
                 int acc = 0;
-                for (int k = 0; k < size; k++) {
+                for (long k = 0; k < size; k++) {
                     acc += a.getData().getAtIndex(ValueLayout.JAVA_INT, baseOffset + k * strideA);
                 }
                 resArray.getData().setAtIndex(ValueLayout.JAVA_INT, offsetRes, acc);
@@ -185,7 +185,7 @@ public class ReduceOps {
     }
 
     public static NDArray sumDoubleAxis(NDArray a,int axis,NDArray resArray){
-        int size= a.internalShapeUnsafe()[axis];
+        long size= a.internalShapeUnsafe()[axis];
         long strideA= a.internalStridesUnsafe()[axis];
 
         for(int i=0;i<resArray.getSize();i++){
@@ -203,8 +203,8 @@ public class ReduceOps {
 
             if(strideA==1){
                 var vAcc = DoubleVector.zero(SPECIESDB);
-                int k = 0;
-                int loopbound = SPECIESDB.loopBound(size);
+                long k = 0;
+                long loopbound = SPECIESDB.loopBound(size);
                 for(; k<loopbound ;k += DB_VL){
                     var vVal = DoubleVector.fromMemorySegment(SPECIESDB, a.getData(), (baseOffset + k) * 8L, ORDER);
                     vAcc = vAcc.add(vVal);
@@ -216,7 +216,7 @@ public class ReduceOps {
                 resArray.getData().setAtIndex(ValueLayout.JAVA_DOUBLE, offsetRes, acc);
             }else{
                 double acc = 0f;
-                for (int k = 0; k < size; k++) {
+                for (long k = 0; k < size; k++) {
                     acc += a.getData().getAtIndex(ValueLayout.JAVA_DOUBLE, baseOffset + k * strideA);
                 }
                 resArray.getData().setAtIndex(ValueLayout.JAVA_DOUBLE, offsetRes, acc);
@@ -254,7 +254,7 @@ public class ReduceOps {
     }
 
     public static NDArray maxFloatAxis(NDArray a,int axis,NDArray resArray){
-        int size= a.internalShapeUnsafe()[axis];
+        long size= a.internalShapeUnsafe()[axis];
         long strideA= a.internalStridesUnsafe()[axis];
 
         for(int i=0;i<resArray.getSize();i++){
@@ -272,8 +272,8 @@ public class ReduceOps {
 
             if(strideA==1){
                 var vAcc = FloatVector.broadcast(SPECIES, Float.NEGATIVE_INFINITY);
-                int k = 0;
-                int loopbound = SPECIES.loopBound(size);
+                long k = 0;
+                long loopbound = SPECIES.loopBound(size);
                 for(; k<loopbound ;k += VL){
                     var vVal = FloatVector.fromMemorySegment(SPECIES, a.getData(), (baseOffset + k) * 4L, ORDER);
                     vAcc = vAcc.max(vVal);
@@ -285,7 +285,7 @@ public class ReduceOps {
                 resArray.getData().setAtIndex(ValueLayout.JAVA_FLOAT, offsetRes, acc);
             }else{
                 float acc = Float.NEGATIVE_INFINITY;
-                for (int k = 0; k < size; k++) {
+                for (long k = 0; k < size; k++) {
                     acc = Math.max(acc, a.getData().getAtIndex(ValueLayout.JAVA_FLOAT, baseOffset + k * strideA));
                 }
                 resArray.getData().setAtIndex(ValueLayout.JAVA_FLOAT, offsetRes, acc);
@@ -321,7 +321,7 @@ public class ReduceOps {
     }
 
     public static NDArray maxIntAxis(NDArray a,int axis,NDArray resArray){
-        int size= a.internalShapeUnsafe()[axis];
+        long size= a.internalShapeUnsafe()[axis];
         long strideA= a.internalStridesUnsafe()[axis];
 
         for(int i=0;i<resArray.getSize();i++){
@@ -339,8 +339,8 @@ public class ReduceOps {
 
             if(strideA==1){
                 var vAcc = IntVector.broadcast(SPECIESINT, Integer.MIN_VALUE);
-                int k = 0;
-                int loopbound = SPECIESINT.loopBound(size);
+                long k = 0;
+                long loopbound = SPECIESINT.loopBound(size);
                 for(; k<loopbound ;k += INT_VL){
                     var vVal = IntVector.fromMemorySegment(SPECIESINT, a.getData(), (baseOffset + k) * 4L, ORDER);
                     vAcc = vAcc.max(vVal);
@@ -352,7 +352,7 @@ public class ReduceOps {
                 resArray.getData().setAtIndex(ValueLayout.JAVA_INT, offsetRes, acc);
             }else{
                 int acc = Integer.MIN_VALUE;
-                for (int k = 0; k < size; k++) {
+                for (long k = 0; k < size; k++) {
                     acc = Math.max(acc, a.getData().getAtIndex(ValueLayout.JAVA_INT, baseOffset + k * strideA));
                 }
                 resArray.getData().setAtIndex(ValueLayout.JAVA_INT, offsetRes, acc);
@@ -388,7 +388,7 @@ public class ReduceOps {
     }
 
     public static NDArray maxDoubleAxis(NDArray a,int axis,NDArray resArray){
-        int size= a.internalShapeUnsafe()[axis];
+        long size= a.internalShapeUnsafe()[axis];
         long strideA= a.internalStridesUnsafe()[axis];
 
         for(int i=0;i<resArray.getSize();i++){
@@ -406,8 +406,8 @@ public class ReduceOps {
 
             if(strideA==1){
                 var vAcc = DoubleVector.broadcast(SPECIESDB, Double.NEGATIVE_INFINITY);
-                int k = 0;
-                int loopbound = SPECIESDB.loopBound(size);
+                long k = 0;
+                long loopbound = SPECIESDB.loopBound(size);
                 for(; k<loopbound ;k += DB_VL){
                     var vVal = DoubleVector.fromMemorySegment(SPECIESDB, a.getData(), (baseOffset + k) * 8L, ORDER);
                     vAcc = vAcc.max(vVal);
@@ -419,7 +419,7 @@ public class ReduceOps {
                 resArray.getData().setAtIndex(ValueLayout.JAVA_DOUBLE, offsetRes, acc);
             }else{
                 double acc = Double.NEGATIVE_INFINITY;
-                for (int k = 0; k < size; k++) {
+                for (long k = 0; k < size; k++) {
                     acc = Math.max(acc, a.getData().getAtIndex(ValueLayout.JAVA_DOUBLE, baseOffset + k * strideA));
                 }
                 resArray.getData().setAtIndex(ValueLayout.JAVA_DOUBLE, offsetRes, acc);
@@ -457,7 +457,7 @@ public class ReduceOps {
     }
 
     public static NDArray minFloatAxis(NDArray a,int axis,NDArray resArray){
-        int size= a.internalShapeUnsafe()[axis];
+        long size= a.internalShapeUnsafe()[axis];
         long strideA= a.internalStridesUnsafe()[axis];
 
         for(int i=0;i<resArray.getSize();i++){
@@ -475,8 +475,8 @@ public class ReduceOps {
 
             if(strideA==1){
                 var vAcc = FloatVector.broadcast(SPECIES, Float.POSITIVE_INFINITY);
-                int k = 0;
-                int loopbound = SPECIES.loopBound(size);
+                long k = 0;
+                long loopbound = SPECIES.loopBound(size);
                 for(; k<loopbound ;k += VL){
                     var vVal = FloatVector.fromMemorySegment(SPECIES, a.getData(), (baseOffset + k) * 4L, ORDER);
                     vAcc = vAcc.min(vVal);
@@ -488,7 +488,7 @@ public class ReduceOps {
                 resArray.getData().setAtIndex(ValueLayout.JAVA_FLOAT, offsetRes, acc);
             }else{
                 float acc = Float.POSITIVE_INFINITY;
-                for (int k = 0; k < size; k++) {
+                for (long k = 0; k < size; k++) {
                     acc = Math.min(acc, a.getData().getAtIndex(ValueLayout.JAVA_FLOAT, baseOffset + k * strideA));
                 }
                 resArray.getData().setAtIndex(ValueLayout.JAVA_FLOAT, offsetRes, acc);
@@ -524,7 +524,7 @@ public class ReduceOps {
     }
 
     public static NDArray minIntAxis(NDArray a,int axis,NDArray resArray){
-        int size= a.internalShapeUnsafe()[axis];
+        long size= a.internalShapeUnsafe()[axis];
         long strideA= a.internalStridesUnsafe()[axis];
 
         for(int i=0;i<resArray.getSize();i++){
@@ -542,8 +542,8 @@ public class ReduceOps {
 
             if(strideA==1){
                 var vAcc = IntVector.broadcast(SPECIESINT, Integer.MAX_VALUE);
-                int k = 0;
-                int loopbound = SPECIESINT.loopBound(size);
+                long k = 0;
+                long loopbound = SPECIESINT.loopBound(size);
                 for(; k<loopbound ;k += INT_VL){
                     var vVal = IntVector.fromMemorySegment(SPECIESINT, a.getData(), (baseOffset + k) * 4L, ORDER);
                     vAcc = vAcc.min(vVal);
@@ -555,7 +555,7 @@ public class ReduceOps {
                 resArray.getData().setAtIndex(ValueLayout.JAVA_INT, offsetRes, acc);
             }else{
                 int acc = Integer.MAX_VALUE;
-                for (int k = 0; k < size; k++) {
+                for (long k = 0; k < size; k++) {
                     acc = Math.min(acc, a.getData().getAtIndex(ValueLayout.JAVA_INT, baseOffset + k * strideA));
                 }
                 resArray.getData().setAtIndex(ValueLayout.JAVA_INT, offsetRes, acc);
@@ -591,7 +591,7 @@ public class ReduceOps {
     }
 
     public static NDArray minDoubleAxis(NDArray a,int axis,NDArray resArray){
-        int size= a.internalShapeUnsafe()[axis];
+        long size= a.internalShapeUnsafe()[axis];
         long strideA= a.internalStridesUnsafe()[axis];
 
         for(int i=0;i<resArray.getSize();i++){
@@ -609,8 +609,8 @@ public class ReduceOps {
 
             if(strideA==1){
                 var vAcc = DoubleVector.broadcast(SPECIESDB, Double.POSITIVE_INFINITY);
-                int k = 0;
-                int loopbound = SPECIESDB.loopBound(size);
+                long k = 0;
+                long loopbound = SPECIESDB.loopBound(size);
                 for(; k<loopbound ;k += DB_VL){
                     var vVal = DoubleVector.fromMemorySegment(SPECIESDB, a.getData(), (baseOffset + k) * 8L, ORDER);
                     vAcc = vAcc.min(vVal);
@@ -622,7 +622,7 @@ public class ReduceOps {
                 resArray.getData().setAtIndex(ValueLayout.JAVA_DOUBLE, offsetRes, acc);
             }else{
                 double acc = Double.POSITIVE_INFINITY;
-                for (int k = 0; k < size; k++) {
+                for (long k = 0; k < size; k++) {
                     acc = Math.min(acc, a.getData().getAtIndex(ValueLayout.JAVA_DOUBLE, baseOffset + k * strideA));
                 }
                 resArray.getData().setAtIndex(ValueLayout.JAVA_DOUBLE, offsetRes, acc);
