+++ src/main/java/jnum/jnumutils/ShapeUtil.java
@@ -1,14 +1,16 @@
 package jnum.jnumutils;
 
+import java.lang.foreign.MemoryLayout;
+import java.lang.foreign.MemorySegment;
 import java.util.Arrays;
 
 import jnum.DType;
 
 public class ShapeUtil {
 
-    public static int[] calculateDefaultStrides(int[] shape){
-        int[] strides=new int[shape.length];
-        int currentstride=1;
+    public static long[] calculateDefaultStrides(long[] shape){
+        long[] strides=new long[shape.length];
+        long currentstride=1;
         for(int i=shape.length-1;i>=0;i--){
             strides[i]=currentstride;
             currentstride*=shape[i];
@@ -16,12 +18,12 @@ public class ShapeUtil {
         return strides;
     }
 
-    public static int[] calculateBroadcastShape(int[] shapeA, int[] shapeB){
+    public static long[] calculateBroadcastShape(long[] shapeA, long[] shapeB){
         int maxDims=Math.max(shapeA.length, shapeB.length);
-        int[] result=new int[maxDims];
+        long[] result=new long[maxDims];
         for (int i = 1; i <= maxDims; i++) {
-            int dimA = (shapeA.length - i >= 0) ? shapeA[shapeA.length - i] : 1;
-            int dimB = (shapeB.length - i >= 0) ? shapeB[shapeB.length - i] : 1;
+            long dimA = (shapeA.length - i >= 0) ? shapeA[shapeA.length - i] : 1;
+            long dimB = (shapeB.length - i >= 0) ? shapeB[shapeB.length - i] : 1;
 
             if (dimA == dimB) {
                 result[maxDims - i] = dimA;
@@ -36,11 +38,11 @@ public class ShapeUtil {
         return result;
     }
 
-    public static int[] calculateReductionShape(int[] shape, int axis) {
+    public static long[] calculateReductionShape(long[] shape, int axis) {
         if (axis < 0 || axis >= shape.length) {
             throw new IllegalArgumentException("Axis " + axis + " is out of bounds for shape " + Arrays.toString(shape));
         }
-        int[] reducedShape = new int[shape.length - 1];
+        long[] reducedShape = new long[shape.length - 1];
         for (int i = 0, j = 0; i < shape.length; i++) {
             if (i != axis) {
                 reducedShape[j++] = shape[i];
@@ -49,11 +51,13 @@ public class ShapeUtil {
         return reducedShape;
     }
 
-    public static long getByteOffset(int[] coords, int[] strides, DType dtype) {
+    public static long getByteOffset(long[] coords, long[] strides, DType dtype) {
         long elementOffset = 0;
         for (int i = 0; i < coords.length; i++) {
-            elementOffset += (long) coords[i] * strides[i];
+            elementOffset += coords[i] * strides[i];
         }
         return elementOffset * dtype.layout.byteSize();
     }
+
+
 }
